requirements.txt
openai>=1.0.0
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
